🩹 [Patch]: Enable verbose and debug output preferences in action script (#80)
## Description

This pull request includes updates to the verbosity and debug
preferences for test outputs in the `README.md` and `action.yml` files.
The most important changes include adding new input parameters for
`VerbosePreference` and `DebugPreference` and updating the run commands
to utilize these new parameters.

Documentation updates:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R106-R107):
Added descriptions for `VerbosePreference` and `DebugPreference` options
to the list of job inputs.

Configuration updates:

*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6R30-R37):
Added `VerbosePreference` and `DebugPreference` as new input parameters
with descriptions and default values.
*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6R60-R65):
Updated the `runs` section to include the new `VerbosePreference` and
`DebugPreference` parameters and set their values in the run commands.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas