🪲 [Fix]: Fix script path and improve test result handling (#91)
## Description

This pull request includes changes to the GitHub Actions workflow and
the PowerShell script used for testing. The most important changes
include updating the script path in the action configuration and
modifying the handling of test results in the PowerShell script.

### Workflow Updates:

*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L72-R72):
Updated the script path in the `runs:` section to use a more
straightforward syntax.

### Script Enhancements:

*
[`scripts/main.ps1`](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011R53-R63):
Improved the handling of test results by setting GitHub output values
for passed tests and using a return variable to determine the exit code.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas