🪲 [Fix]: Fix conditional logic for test result handling (#92)
## Description

This pull request includes a small change to the `scripts/main.ps1`
file. The change modifies the conditional logic to use an `elseif`
statement instead of a separate `if` statement, ensuring that the script
correctly handles the cases where tests have failed or all tests have
passed.

*
[`scripts/main.ps1`](diffhunk://#diff-dc2e5a659836b1b73abb03421c567f5018c2755677c4a0aa764cb26117b68011L55-R55):
Changed the conditional logic from a separate `if` statement to an
`elseif` statement to handle test results more accurately.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas